tests: include switch for silent mode
Silent mode is a required variable, and tests will fail to import app.py
without it in os.environ. By default we set it to 1, because of the XSRF
failures being printed out, which I find kind of annoying. Of course it
won't break if you set it to 0, so feel free to do that if you like those
messages I guess.