models: change Change.className to be indexed
WARNING: This change makes the current Datastore schema
backwards-incompatible. There is no migration utility as of now, but
I might be forced to create one eventually.

Well, maybe not, because the schema versions are unlikely to change much,
but the changes might be annoying for some users, because now it involves
removing everything from your instance and starting anew. A migration
utility might help solve this, but not too important as of now.