LessonsColumn: fix accidental last item duplication
When there would be no trailing empties, the while() wouldn't be
executed, but the .push would. There were no real test cases to
account for that, and now I've added one (or two), also modifying
the existing one to look not just at the length, but at the whole
array that's returned.