Drop scripts from non-logged in users
While doing this really short task, I came up with an entire essay
about what this change means in the grand Edaemon infrastructure.

Essentially these scripts provide not only the pretty dropdown
which appears when you are logged in and allows you to log out
(which I can really argue is quite useless and should be made into
a link instead, but I digress), but also the responsive navbar
collapse. It seems fine until you consider that there is no way
for users to log in when the script is not loaded. It really has
a huge lot of unnecessary bloat which weighs down the page, so I
do want to get rid of it.

I think I'll just move the login link to the bottom right corner
instead. Solves a few things.

Resolves: #30