Remove almost all of the navbar!
This last trick (using a link with .navbar-toggle for mobile
devices) I came up while working on paulsnar/ytstream, I don't
remember the exact circumstances but the bloat of React and
Immutable.js was getting so huge that I didn't want to load
Bootstrap's external scripts and jQuery or something like that,
and I only had one link to be shown in the navbar. So I went and
misused .navbar-toggle on a link to make it appear on mobile
devices! And also it looks not terrible, I wouldn't recommend it
but it's a fine hack, and for the admin page here we can remove
the scripts entirely, cutting down on a few bytes of bandwidth,
hopefully!

I am very proud of this "hack". Now the navbar is almost
completely useless!